Adding Softmax Player
The softmax algorithm is an alternative to the epsilon-greedy algorithm.
Instead of picking based on a set probability, it uses a 'temperature'
parameter, along with the known data to set the distribution of the
alternatives.

lib/bandit.rb requires the softmax player
lib/bandit/players/base.rb adds a softmax option to the factory
lib/bandit/players/softmax.rb implements the player
players.rdoc provides a short explanation