Adding Rake test_player task
Usage: rake bandit:test_player[<test_name>{,<true|false>}]

The test_player task is called with one mandatory argument, the name of
an experiment, and one optional parameter, a boolean determing whether
or not to force.  The configured player for the given environment is the
player that is tested.

Force is only required if running the test in a non-development,
non-test environment.

The purpose of the task to to check if conversion_rates line up nicely
with the player's options.  It can be used to check the accuracy of
players, or to learn what kind of results they produce.